pharmsol

Module simulator

Modules

equation

fitting

Type Aliases

AnalyticalEq

This closure represents an Analytical solution of the model. See [analytical] module for examples.

DiffEq

This closure represents the differential equation of the model.

Diffusion

This closure represents the diffusion term of a stochastic differential equation model.

Drift

This closure represents the drift term of a stochastic differential equation model.

Fa

This closure represents the fraction absorbed (also called bioavailability or protein binding) of the model.

Init

This closure represents the initial state of the system.

Lag

This closure represents the lag time of the model.

M

Type alias for matrix type used in simulations

Neqs

The number of states and output equations of the model.

Out

This closure represents the output equation of the model.

SecEq

This closure represents the secondary equation of the model.

T

Type alias for floating point values used in simulations

V

Type alias for state vector type used in simulations