pharmsol::simulator::equation::analytical::three_compartment_models

Function three_compartments_with_absorption

pub fn three_compartments_with_absorption(
    x: &V,
    p: &V,
    t: T,
    rateiv: V,
    _cov: &Covariates,
) -> V

Analytical solution for three compartment model with first-order absorption. Assumptions:

p is a vector of length 6 with ka, k10, k12, k13, k21 and k31 in that order. ka represents the absorption rate constant k10 represents the elimination rate constant from compartment 1 (central). The remaining parameters of the form kij represent the transfer rate constants from the ith to the jth compartment

rateiv is a vector of length 1 with the value of the infusion rate (only one drug)

x is a vector of length 4

covariates are not used