pharmsol::simulator::fitting

Trait OptimalSupportPoint

pub trait OptimalSupportPoint<E: Equation> {
    // Required method
    fn optimal_support_point(
        &self,
        equation: &E,
        point: &Array1<f64>,
    ) -> Array1<f64>;
}

Trait for finding the optimal support point for a subject

Required Methods

fn optimal_support_point( &self, equation: &E, point: &Array1<f64>, ) -> Array1<f64>

Finds the optimal support point for a subject

Arguments

equation - Reference to the equation model to use for simulation

point - Initial parameter values to start optimization from

Returns

Optimized parameter values as an ndarray Array1