pharmsol::simulator
Type Alias AnalyticalEq

pub type AnalyticalEq = fn(_: &V, _: &V, _: T, _: V, _: &Covariates) -> V;

This closure represents an Analytical solution of the model. See [analytical] module for examples.

Parameters

x: The state vector at time t

p: The parameters of the model; Use the [fetch_params!] macro to extract the parameters

t: The time at which the output equation is evaluated

rateiv: A vector of infusion rates at time t

cov: A reference to the covariates at time t; Use the [fetch_cov!] macro to extract the covariates

TODO: Remove covariates. They are not used in the analytical solution