pharmsol::simulator
Type Alias Neqs

pub type Neqs = (usize, usize);

The number of states and output equations of the model The first element is the number of states and the second element is the number of output equations This is used to initialize the state vector and the output vector Example:

let neqs = (2, 1);

This means that the system of equations has 2 states and there is only 1 output equation.