pharmsol::simulator
Type Alias Out

pub type Out = fn(_: &DVector<f64>, _: &DVector<f64>, _: f64, _: &Covariates, _: &mut DVector<f64>);

This closure represents the output equation of the model: Params:

x: The state vector at time t

p: The parameters of the model; Use the [fetch_params!] macro to extract the parameters

t: The time at which the output equation is evaluated

cov: A reference to the covariates at time t; Use the [fetch_cov!] macro to extract the covariates

y: A mutable reference to the output vector at time t Example:

use pharmsol::*;
let out = |x, p, t, cov, y| {
  fetch_params!(p, ka, ke, v);
  y[0] = x[1] / v;
};