Module pharmsol::simulator

Modules

fitting

Enums

Equation

Functions

get_population_predictions

Type Aliases

AnalyticalEq

This closure represents an Analytical solution of the model, see [analytical] module for examples. Params:

DiffEq

This closure represents the differential equation of the model: Params:

Diffusion

This closure represents the diffusion term of the model: Params:

Drift

This closure represents the drift term of the model: Params:

Fa

This closure represents the fraction absorbed (also called bioavailability or protein binding) of the model, the fa term is used to adjust the amount of drug that is absorbed into the system. Params:

Init

This closure represents the initial state of the system: Params:

Lag

This closure represents the lag time of the model, the lag term delays the only the boluses going into an specific comparment. Params:

Neqs

The number of states and output equations of the model The first element is the number of states and the second element is the number of output equations This is used to initialize the state vector and the output vector Example:

Out

This closure represents the output equation of the model: Params:

SecEq

This closure represents the secondary equation of the model, secondary equations are used to update the parameter values based on the covariates. Params: